Moved stream functions into LinearVortexSource
